Conflicts:
	IF/Resources/Battle/Battle_soldier.plist
	IF/Resources/Imperial/Imperial_22.plist
	IF/Resources/Imperial/Imperial_22.pvr.ccz

+++ IF/Classes/Ext/CCMathUtils.cpp
@@ -70,7 +70,7 @@ float CCMathUtils::getAngle(CCPoint from, CCPoint to)
 	return result;
 }
 
-int CCMathUtils::getCurrentTime(){
+long CCMathUtils::getCurrentTime(){
     struct timeval now;
     
     gettimeofday(&now, nullptr);
